backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
psycopg[binary]==3.2.3
celery==5.4.0
redis==5.2.0
dj-database-url==2.3.0
httpx==0.27.2


